utente può modificare solo progetti di cui è leader

+++ app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
 use App\Models\Project;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
@@ -15,22 +16,33 @@ class ProjectController extends Controller
     }
 
     public function create(){
-        return view('progetti.create');
+        $members=User::all();
+        return view('progetti.create',compact('members'));
     }
     public function store(Request $request){
 
-        $validated = $request->validate(['title' => ['required','min:3'],'description' => ['min:3'],'status' => ['min:3']]);
+        $validated = $request->validate(['title' => ['required','min:3'],'description' => ['min:3'],'status' => ['min:3'],'leaderemail' => ['min:3','required']]);
 
         $title = request()->validate([
             'title' => ['required','min:3']
         ]);
 
-        $exists = Project::where('title', $title)->exists();
+        $leadermail = request()->validate([
+            'leaderemail' => ['required','min:3']
+        ]);
+        $existsTitle = Project::where('title', $title)->exists();
+        $existsEmail = User::where('email', $leadermail)->exists();
 
-        if($exists){
+        if($existsTitle){
 
             throw ValidationException::withMessages([
                 'title' => 'Questo progetto è già presente'
+            ]);
+        }
+        if(!$existsEmail){
+
+            throw ValidationException::withMessages([
+                'leaderemail' => 'Questo utente non esiste'
             ]);
         }
         Project::create($validated);
@@ -38,11 +50,16 @@ class ProjectController extends Controller
     }
 
     public function edit(Project $project){
+            if (Auth::user()->email != $project->leaderemail) {
+            abort(403, 'Accesso non autorizzato');
+        }
         $members=User::all();
         return view('progetti.edit', compact('project','members'));
     }
     public function update(Request $request,Project $project){
-
+        if (Auth::user()->email != $project->leaderemail) {
+            abort(403, 'Accesso non autorizzato');
+        }
         $validated = $request->validate(['title' => ['required','min:3'],'description' => ['required','min:3'],'status' => ['required']]);
 
         $project->update($validated);
@@ -50,13 +67,17 @@ class ProjectController extends Controller
         return to_route('progetti.index');
     }
     public function destroy(Project $project){
-
+        if (Auth::user()->email != $project->leaderemail) {
+            abort(403, 'Accesso non autorizzato');
+        }
         $project->delete();
 
         return back()->with('message','Course destroyed');
     }
     public function assignMember(Request $request, Project $project){
-
+        if (Auth::user()->email != $project->leaderemail) {
+            abort(403, 'Accesso non autorizzato');
+        }
         if ($project->users->contains($request->member)){
             return back()->with('message','Course already assigned');
         }
@@ -64,7 +85,9 @@ class ProjectController extends Controller
         return back()->with('message','Course assigned');
     }
     public function revokeMember(Project $project, User $member){
-
+        if (Auth::user()->email != $project->leaderemail) {
+            abort(403, 'Accesso non autorizzato');
+        }
         if ($project->users->contains($member->id)){
             $project->users()->detach($member->id);
             return back()->with('message','Course removed');
